requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-autocomplete-light==3.9.4
django-crispy-forms==2.0
idna==3.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
